Refactor movement handling keybindings to be done entirely in one file.
This is a fairly substantial refactoring to ensure that all
text-editing keybindings are handled in text-buffer.ts rather than
split across that file and InputPrompt.tsx.

One minor consequence is that we had to remove ctrl-e as a way of
invoking an external editor from multiline-editor.tsx, as it conflicts
with the ctrl-e movement to end of line. It still can be invoked with
ctrl+x.

Will follow-up with a few other cleanups in later commits.