Refactor: Add GeminiRespondingSpinner to make use of streamingState more idiomatic.
This Spinner should be used everywhere in the UI where we want to use a spinner
to avoid flicker caused by showing spinners when we are not responding.
- If `streamingState` is `Responding`, the component renders a spinner.
- If `nonRespondingDisplay` is  set it days the fallback content.
- Otherwise (not `Responding` and no `nonRespondingDisplay`), it renders `null`.

Usage in `LoadingIndicator.tsx` and `ToolMessage.tsx` has been updated to pass the appropriate `nonRespondingDisplay` string to maintain previous visual behavior. Tests for `LoadingIndicator.test.tsx` and `ToolMessage.test.tsx` have been updated accordingly.
Fix bug where the default display for the Spinner in ToolMessage wasn't aligned with the Spinner type.