feat: Use concurrently to run GCP start script
Introduces the `concurrently` package to simplify the GCP development startup process.

Previously, developers had to run the telemetry script and the main application start script in separate terminals. This change updates the `start:gcp` script to use `concurrently`, allowing both processes to be launched and managed with a single command. This improves the developer experience and reduces the chance of forgetting to start one of the required processes.